TollCalculator: Moves default specifications to separate package
